app-engine/requirements.txt
Django==1.7
django-grappelli==2.6.1
PyJWT==0.2.3
